requirements.txt
anyio==3.6.2
asyncio==3.4.3
beanie==1.13.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.1
h11==0.14.0
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
motor==3.1.1
multidict==6.0.2
orjson==3.8.0
pydantic==1.10.2
pymongo==4.3.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
toml==0.10.2
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.0
websockets==10.4
yarl==1.8.1
